refactor: extract shared verifyTargetUserAccess helper
Addresses Adma's code review feedback to reduce duplication between
RunService and RunEventService authorization logic.
